requirements.txt
appdirs==1.4.0
Django>=1.10.5
packaging==16.8
pyparsing==2.1.10
six==1.10.0
bleach>=2.0.0
coverage>=4.4.1
dj-database-url>=0.4.2
dj-static>=0.0.6
django-autoslug>=1.9
django-taggit>=0.22
gunicorn>=19.7
Markdown>=2.6
Pillow>=4.2
psycopg2>=2.7
python-decouple>=3
pytz>=2017.2
Unipath>=1.1
whitenoise>=3.3.0
django_extensions
Faker
django-crispy-forms
aloe
pandas
validators
csvvalidator
